Last fix of solenoid state and temperature conversion.

+++ heater_on_oil/src/temperature/temperature.h
@@ -5,7 +5,7 @@
 
 float Temperature_GetAcc(void);
 uint16_t Temperature_GetRawRef(void);
-uint16_t Temperature_GetRef(void);
+uint8_t Temperature_GetRef(void);
 void Temperature_Init(void);
 void Temperature_Cyclic(void);
